add reusable fetchRelevantRows
