+++ express/scripts/utils/relevant.js
@@ -1,0 +1,30 @@
+import { getLibs } from '../utils.js';
+
+const { getConfig } = await import(`${getLibs()}/utils/utils.js`);
+
+// eslint-disable-next-line import/prefer-default-export
+export async function fetchRelevantRows(path) {
+  if (!window.relevantRows) {
+    try {
+      const { prefix } = getConfig().locale;
+      const resp = await fetch(`${prefix}/express/relevant-rows.json`);
+      window.relevantRows = resp.ok ? (await resp.json()).data : [];
+    } catch {
+      const resp = await fetch('/express/relevant-rows.json');
+      window.relevantRows = resp.ok ? (await resp.json()).data : [];
+    }
+  }
+
+  if (window.relevantRows.length) {
+    const relevantRow = window.relevantRows.find((p) => path === p.path);
+    const { env } = getConfig();
+
+    if (env && env.name === 'stage') {
+      return relevantRow || null;
+    }
+
+    return relevantRow && relevantRow.live !== 'N' ? relevantRow : null;
+  }
+
+  return null;
+}
